Bot/requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b8
aiohttp==3.8.4
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.26.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.6.0
idna==3.10
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
typing_extensions==4.13.2
yarl==1.20.0
